+++ examples/StreetlightsAPI/API.cs
@@ -1,7 +1,10 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Text.Json;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
+using Saunter.Attributes;
 
 namespace StreetlightsAPI
 {
@@ -16,36 +19,67 @@ namespace StreetlightsAPI
         /// Lat-Long coordinates of the streetlight.
         /// </summary>
         public double[] Position { get; set; }
+
+        /// <summary>
+        /// History of light intensity measurements
+        /// </summary>
+        public List<KeyValuePair<DateTime, int>> LightIntensity { get; set; }
     }
-    
+
+    public class LightMeasuredEvent
+    {
+        /// <summary>
+        /// Id of the streetlight.
+        /// </summary>
+        public int Id { get; set; }
+
+        /// <summary>
+        /// Light intensity measured in lumens.
+        /// </summary>
+        public int Lumens { get; set; }
+
+        /// <summary>
+        /// Light intensity measured in lumens.
+        /// </summary>
+        public DateTime SentAt { get; set; }
+    }
+
+    [AsyncApi]
     [ApiController]
-    [Route("api/streetlights")]
+    [Route("")]
     public class StreetlightsController
     {
+        private const string PublishLightMeasuredTopic = "publish/light/measured";
+
+
         // Simulate a database of streetlights
         private static int StreetlightSeq = 2;
         private static readonly List<Streetlight> StreetlightDatabase = new List<Streetlight>
         {
-            new Streetlight { Id = 1, Position = new [] { -36.320320, 175.485986 } },
+            new Streetlight { Id = 1, Position = new [] { -36.320320, 175.485986 }, LightIntensity = new() },
         };
         
         private readonly IStreetlightMessageBus _streetlightMessageBus;
+        private readonly ILogger _logger;
 
-        public StreetlightsController(IStreetlightMessageBus streetlightMessageBus)
+        public StreetlightsController(IStreetlightMessageBus streetlightMessageBus, ILoggerFactory loggerFactory)
         {
             _streetlightMessageBus = streetlightMessageBus;
+            _logger = loggerFactory.CreateLogger<StreetlightsController>();
         }
         
         /// <summary>
         /// Get all streetlights
         /// </summary>
         [HttpGet]
+        [Route("api/streetlights")]
         public IEnumerable<Streetlight> Get() => StreetlightDatabase;
 
         /// <summary>
         /// Add a new streetlight
         /// </summary>
         [HttpPost]
+        [Route("api/streetlights")]
         public Streetlight Add([FromBody] Streetlight streetlight)
         {
             streetlight.Id = StreetlightSeq++;
@@ -54,16 +88,28 @@ namespace StreetlightsAPI
         }
 
         /// <summary>
-        /// Measure environmental lighting conditions for a particular streetlight.
+        /// Inform about environmental lighting conditions for a particular streetlight.
         /// </summary>
+        [Channel(PublishLightMeasuredTopic, Servers = new []{"webapi"})]
+        [PublishOperation(typeof(LightMeasuredEvent))]
         [HttpPost]
-        [Route("{id}/measure-light")]
-        public void MeasureLight([FromRoute] int id)
+        [Route(PublishLightMeasuredTopic)]
+        public void MeasureLight([FromBody] LightMeasuredEvent lightMeasuredEvent)
         {
-            var streetlight = StreetlightDatabase.Single(s => s.Id == id);
-            var lumens = new Random().Next(0, 3000); // Simulate "measuring" the light intensity
-            
-            _streetlightMessageBus.PublishLightMeasuredEvent(streetlight, lumens);
+            lightMeasuredEvent.SentAt = DateTime.Now;
+
+            var payload = JsonSerializer.Serialize(lightMeasuredEvent);
+
+            _logger.LogInformation("Received message on {Topic} with payload {Payload} ", PublishLightMeasuredTopic, payload);
+
+            var streetlight = StreetlightDatabase.SingleOrDefault(s => s.Id == lightMeasuredEvent.Id);
+            if (streetlight != null)
+            {
+                streetlight.LightIntensity.Add(new(lightMeasuredEvent.SentAt, lightMeasuredEvent.Lumens));
+
+                // Re-publish messages we receive
+                _streetlightMessageBus.PublishLightMeasurement(lightMeasuredEvent);
+            }
         }
     }
 }